WIP: Add Javascript modules
