+++ Distribution/Default/package.json
@@ -67,6 +67,7 @@
     "scss": "yarn gulp scss",
     "lint": "yarn gulp lint",
     "js": "yarn gulp js",
+    "mjs": "yarn gulp mjs",
     "optimizeImages": "yarn gulp optimizeImages",
     "optimizeSvg": "yarn gulp optimizeSvg",
     "compress": "yarn gulp compress",
@@ -75,12 +76,12 @@
   "dependencies": {},
   "devDependencies": {
     "ansi-colors": "^1.1.0",
-    "autoprefixer": "^8.2.0",
+    "autoprefixer": "^8.3.0",
     "babel-cli": "^6.26.0",
     "babel-plugin-external-helpers": "^6.22.0",
     "babel-preset-env": "^1.6.1",
     "browser-sync": "2.23.6",
-    "caniuse-db": "^1.0.30000827",
+    "caniuse-db": "^1.0.30000830",
     "css-mqpacker": "^6.0.2",
     "cssnano": "^3.10.0",
     "del": "^3.0.0",
@@ -105,7 +106,7 @@
     "gulp-plumber": "^1.2.0",
     "gulp-postcss": "^7.0.1",
     "gulp-rename": "^1.2.2",
-    "gulp-rollup": "^2.16.2",
+    "gulp-rollup-each": "^2.1.0",
     "gulp-sass": "^4.0.1",
     "gulp-size": "^3.0.0",
     "gulp-sourcemaps": "^2.6.4",
@@ -136,7 +137,7 @@
     "postcss-selector-not": "^3.0.1",
     "postcss-short": "^4.1.0",
     "postcss-vmax": "^1.0.0",
-    "prettier": "^1.11.1",
+    "prettier": "^1.12.0",
     "prettyjson": "^1.2.1",
     "rollup": "^0.57.1",
     "rollup-plugin-amd": "^3.0.0",
