+++ tasks/mjs.js
@@ -1,0 +1,7 @@
+"use strict";
+
+if (!config.tasks.mjs) {
+    return false;
+}
+
+module.exports = require("./jsRenderShared")("mjs");
