+++ tasks/mjsLint.js
@@ -1,0 +1,7 @@
+"use strict";
+
+if (!config.tasks.jsLint || !config.tasks.mjs) {
+    return false;
+}
+
+module.exports = require("./jsLintShared")("mjs");
